+++ print_funcs1.c
@@ -3,11 +3,13 @@
 /**
  * print_unsigned_int - prints an unsigned integer
  * @ap: arg.
+ * @flag: the flag character.
+ *
  * Return: the count of printed characters.
  */
 int print_unsigned_int(va_list ap, char flag)
 {
-	unsigned long int num = va_arg(ap, unsigned long int);
+	unsigned int num = va_arg(ap, unsigned int);
 	char *str = _itoa(num, 10);
 	int count = 0;
 	(void)flag;
@@ -17,14 +19,17 @@ int print_unsigned_int(va_list ap, char flag)
 	free(str);
 	return (count);
 }
+
 /**
  * print_octal - prints an unsigned integer in octal format
  * @ap: arg.
+ * @flag: the flag character.
+ *
  * Return: the count of printed characters.
  */
 int print_octal(va_list ap, char flag)
 {
-	unsigned long int num = va_arg(ap, unsigned long int);
+	unsigned int num = va_arg(ap, unsigned int);
 	char *str = _itoa(num, 8);
 	int count = 0;
 
@@ -35,15 +40,18 @@ int print_octal(va_list ap, char flag)
 	free(str);
 	return (count);
 }
+
 /**
  * print_hex - prints an unsigned integer in lowercase hexadecimal
  * @ap: arg.
+ * @flag: the flag character.
  *
  * Return: the count of printed characters.
  */
+
 int print_hex(va_list ap, char flag)
 {
-	unsigned long int num = va_arg(ap, unsigned long int);
+	unsigned int num = va_arg(ap, unsigned int);
 	char *str = _itoa(num, 16);
 	int count = 0;
 
@@ -57,12 +65,14 @@ int print_hex(va_list ap, char flag)
 /**
  * print_hex_uppercase - prints an unsigned integer in uppercase hexadecimal
  * @ap: arg.
+ * @flag: the flag character.
  *
  * Return: the count of printed characters.
  */
+
 int print_hex_uppercase(va_list ap, char flag)
 {
-	unsigned long int num = va_arg(ap, unsigned long int);
+	unsigned int num = va_arg(ap, unsigned int);
 	char *str = _itoa(num, 16);
 	int count = 0;
 
@@ -74,10 +84,12 @@ int print_hex_uppercase(va_list ap, char flag)
 	free(str);
 	return (count);
 }
+
 /**
  * print_str_wide - function that prints @ap while converting non-printable
  * characters to their hexadecimal ASCII representation.
  * @ap: the string to be printed.
+ * @flag: the flag character.
  *
  * Return: the count of printed characters.
  */
